🎨  Namespaced all our custom Swan cache
