+++ src/moderation/ModerationData.ts
@@ -53,7 +53,7 @@ class ModerationData {
       this.channel = argument.channel;
     } else if (argument instanceof AkairoClient) {
       this.client = argument;
-      this.channel = this.client.cachedChannels.log;
+      this.channel = this.client.cache.channels.log;
       this.guild = this.channel.guild;
       this.moderator = this.guild.me;
     } else {
